Fixes after merge with dev
