+++ src/api/deviceListApi.js
@@ -4,7 +4,7 @@ let listDevices = [...devices];
 
 export default class DeviceListApi {
   static getDevices () {
-    return fetch('http://localhost:3000/devices')
+    return fetch('http://localhost:3001/devices')
     .then(response => {
       return response.json();
     });
@@ -20,13 +20,13 @@ export default class DeviceListApi {
     });
   }
   static getDevice (id) {
-    return fetch('http://localhost:3000/devices/device/' + id)
+    return fetch('http://localhost:3001/devices/device/' + id)
     .then(response => {
       return response.json();
     });
   }
   static deleteDevice (id) {
-    return fetch('http://localhost:3000/devices/delete/' + id)
+    return fetch('http://localhost:3001/devices/delete/' + id)
       .then(response => {
         return response.text();
       });
