+++ src/server/models/device.js
@@ -1,11 +1,12 @@
 const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
-const devicesSchema = new Schema({
-  id: Number,
-  name: String,
-  items: { type : Array , 'default' : [] },
-  status: Boolean,
-  location: String
+
+const deviceSchema = new Schema({
+    id: Number,
+    name: String,
+    location: String,
+    status: Boolean,
+    items:{ type : Array , "default" : [] }
 });
 
-module.exports = mongoose.model('Device', devicesSchema);
+module.exports = mongoose.model('Device', deviceSchema);
