--- src/server/models/devicesModel.js
@@ -1,12 +1,0 @@
-var mongoose = require('mongoose');
-var Schema = mongoose.Schema;
-
-var Device = new Schema({
-    id: Number,
-    name: String,
-    location: String,
-    status: Boolean,
-    items:{ type : Array , "default" : [] }
-});
-
-module.exports = mongoose.model('Device', Device);
